requirements.txt
asgiref==3.3.4
click==7.1.2
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.12.4
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1


